seat info split logic and comp
